Makes triangle links into their own component
