+++ src/components/index.ts
@@ -11,3 +11,4 @@ export { SEO } from "./seo/seo"
 export { SubHeading } from "./subHeading/subHeading"
 export { RightPage } from "./rightPage/rightPage"
 export { TopHeader } from "./topHeader/topHeader"
+export { TriangleLinks } from "./triangleLinks/triangleLinks"
