+++ src/components/navBar/navBar.scss
@@ -51,3 +51,4 @@
   border-bottom: 100px solid transparent;
   border-right: 100px solid $color-accent-light
 }
+
